requirements.txt
Django==1.10
django-filter==0.14.0
django-grappelli==2.8.1
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
feedparser==5.2.1
Markdown==2.6.6
psycopg2==2.6.2
PyJWT==1.4.2
wheel==0.24.0
django-model-utils-2.5.2
